add minor version in api rest route + refactor geoutil
