+++ lib/api/controllers/routerController.js
@@ -68,7 +68,7 @@ module.exports = function RouterController (kuzzle) {
     }
 
     // create and mount a new router for our API
-    this.router.use('/api/v1', api);
+    this.router.use('/api/v1.0', api);
 
     // create and mount a new router for plugins
     this.pluginRouter = new Router();
