+++ lib/api/core/hotelClerk.js
@@ -604,13 +604,14 @@ function cleanUpRooms (roomId) {
     this.kuzzle.dsl.removeRoom(this.rooms[roomId])
       .then(() => {
         this.kuzzle.pluginsManager.trigger('room:remove', roomId);
-        delete this.rooms[roomId];
         deferred.resolve(roomId);
       })
       .catch(error => {
         this.kuzzle.pluginsManager.trigger('log:error', error);
         deferred.reject(error);
-      });
+      })
+      .finally(() => delete this.rooms[roomId]);
+
   }
   else {
     deferred.resolve(roomId);
